Build uniffi with -F clib as workaround
The problem was that clib and uniffi are sharing the same libhyperbee.so
output. But this confuses the makefile for clibe when uniffi builds the
libhyperbee.so without clib. Tried making two diff libs by adding an arg
to uniffi::setup_scafolding("hyperbeeuniffi") (didnt change the .so
output name) and changing the name in
cargo.toml lib to hyperbeec but this changed the name of the thing we
import.